h2base

+++ Inscription_event/src/main/java/com/example/inscription_event/repository/InsRepo.java
@@ -3,5 +3,5 @@ package com.example.inscription_event.repository;
 import com.example.inscription_event.entity.InsEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InsRepo extends JpaRepository<InsEvent, Integer> {
+public interface InsRepo extends JpaRepository<InsEvent, Long> {
 }
